+++ dejafu/Test/DejaFu/Conc/Internal.hs
@@ -29,13 +29,14 @@ import           Data.Monoid                         ((<>))
 import           Data.Sequence                       (Seq, (<|))
 import qualified Data.Sequence                       as Seq
 
-import           Test.DejaFu.Common
 import           Test.DejaFu.Conc.Internal.Common
 import           Test.DejaFu.Conc.Internal.Memory
 import           Test.DejaFu.Conc.Internal.Threading
+import           Test.DejaFu.Internal
 import           Test.DejaFu.Schedule
 import           Test.DejaFu.STM                     (Result(..),
                                                       runTransaction)
+import           Test.DejaFu.Types
 
 --------------------------------------------------------------------------------
 -- * Execution